requirements.txt
discord.py>=2.3.0
python-dotenv>=0.21.0
PyNaCl>=1.5.0
aiohttp
aiosqlite>=0.19.0